refactor: Move schemas to separate folder
