+++ src/schemas/document.ts
@@ -5,7 +5,7 @@ import {
   objectId,
   standardBoolean,
   standardString,
-} from "~/server/api/helpers/custom-zod-helpers";
+} from "~/schemas/helpers/custom-zod-helpers";
 
 export const createDocumentSchema = z.object({
   title: nonEmptyString,
