+++ src/schemas/user.ts
@@ -1,6 +1,6 @@
 import { AccountRoles } from "@prisma/client";
 import { z } from "zod";
-import { objectId } from "~/server/api/helpers/custom-zod-helpers";
+import { objectId } from "~/schemas/helpers/custom-zod-helpers";
 
 export const updateUserRolesSchema = z.object({
   id: objectId,
