+++ src/server/api/routers/committee-member.ts
@@ -1,10 +1,10 @@
 import { z } from "zod";
-import { objectId } from "~/server/api/helpers/custom-zod-helpers";
+import { objectId } from "~/schemas/helpers/custom-zod-helpers";
 import {
   createMemberSchema,
   updateMemberAsActiveSchema,
   updateMemberSchema,
-} from "~/server/api/helpers/schemas/members";
+} from "~/schemas/member";
 import {
   createTRPCRouter,
   organizationManagementProcedure,
