+++ src/server/api/routers/document.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { objectId } from "~/server/api/helpers/custom-zod-helpers";
+import { objectId } from "~/schemas/helpers/custom-zod-helpers";
 import {
   createTRPCRouter,
   documentProcedure,
