+++ src/server/api/routers/old-committee.ts
@@ -2,11 +2,11 @@ import { z } from "zod";
 import {
   objectId,
   standardBoolean,
-} from "~/server/api/helpers/custom-zod-helpers";
+} from "~/schemas/helpers/custom-zod-helpers";
 import {
   createOldCommitteeSchema,
   updateOldCommitteeSchema,
-} from "~/server/api/helpers/schemas/old-committee";
+} from "~/schemas/old-committee";
 import {
   createTRPCRouter,
   protectedProcedure,
